+++ console/src/main/java/org/nervos/ckb/wallet/Constant.java
@@ -1,0 +1,15 @@
+package org.nervos.ckb.wallet;
+
+/**
+ * Created by duanyytop on 2019-01-29.
+ * Copyright © 2018 Nervos Foundation. All rights reserved.
+ */
+public class Constant {
+
+    public static final String NODE_URL = "http://localhost:8114/";
+    public static final long MIN_CELL_CAPACITY = 40;
+    public static final long MIN_UDT_CELL_CAPACITY = 48;
+
+    public static final int VERSION = 0;
+
+}