+++ core/src/main/java/org/nervos/ckb/methods/response/CkbNodeInfo.java
@@ -1,14 +1,15 @@
 package org.nervos.ckb.methods.response;
 
 import org.nervos.ckb.methods.Response;
+import org.nervos.ckb.methods.type.NodeInfo;
 
 /**
  * Created by duanyytop on 2018-12-21.
  * Copyright © 2018 Nervos Foundation. All rights reserved.
  */
-public class CkbNodeId extends Response<String> {
+public class CkbNodeInfo extends Response<NodeInfo> {
 
-    public String getNodeId() {
+    public NodeInfo getNodeInfo() {
         return result;
     }