+++ core/src/main/java/org/nervos/ckb/methods/type/CellInput.java
@@ -2,13 +2,11 @@ package org.nervos.ckb.methods.type;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.util.List;
-
 /**
  * Created by duanyytop on 2019-01-08.
  * Copyright © 2018 Nervos Foundation. All rights reserved.
  */
-public class Input {
+public class CellInput {
 
     /**
      * previous_output : {"hash":"0x0000000000000000000000000000000000000000000000000000000000000000","index":4294967295}
@@ -19,6 +17,13 @@ public class Input {
     public PreviousOutput previousOutput;
     public Script unlock;
 
+    public CellInput(){}
+
+    public CellInput(PreviousOutput previousOutput, Script unlock) {
+        this.previousOutput = previousOutput;
+        this.unlock = unlock;
+    }
+
     public static class PreviousOutput {
         /**
          * hash : 0x0000000000000000000000000000000000000000000000000000000000000000
@@ -27,6 +32,13 @@ public class Input {
 
         public String hash;
         public long index;
+
+        public PreviousOutput(){}
+
+        public PreviousOutput(String hash, long index) {
+            this.hash = hash;
+            this.index = index;
+        }
 
     }