+++ core/src/main/java/org/nervos/ckb/utils/ObjectMapperFactory.java
@@ -1,6 +1,5 @@
 package org.nervos.ckb.utils;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.*;